feat: spawn concurrent list-repo tasks
Concurrently lookup repositories when caling "list_repos". Because I'm
really not an async Rust expert, there's a bunch of unsafe code to allow
for immutable references to Self. Avoiding those would be awesome, but
even with the unsafe trickery, the speed-up is insane:
instead of listing 30 repos with a response time of ~100ms sequentially,
we make 30 responses in parallel, which only takes a couple hundred ms
in total!