feat: add option to git init current dir
Add an option to add the remote repository as remote `origin` to a local
repository in the current directory.
If the current directory is a repository (.git does not exist), this
option will initialize it first.

TODO: Add config option (or get the settings from the remote directly?)
to automatically pull the default branch.