fix: load default colors on create-config command
The create-config command does not load the config file (because it may
not exist and will be overwritten regardless). This caused issues with
the new way we handle colors - colors are loaded from the global config,
but because we haven't loaded it on `create-config`, gritty panics and
fails to create a new config.

Thus, create a new method `log::load_default_colors, which sets the
global `STYLES` variable to the default colors.