fix(gitlab): invalid_token authentication error
The gitlab crate doesn't seem to care about correctly handling errors,
so we have to match for the GitLab API error messages.

Previously, gritty wouldn't return the correct error kind when using
invalid tokens. For example, instead of printing `not authenticated`
when calling `gritty auth status`, gritty would "crash".

To fix this, match for the `invalid_token` string when GitLab returns an
error.